feat: Ask for a secret when user has an account on GitLab
Currently, the fork of a public repo is created when user either
1. Register, entering the secret
2. Logs in through the course web-page
The second option allows one to enroll for a course on Manytask even
if they don't know the secret. The only information needed is the
course URL. This commit adds an extra form, which asks for a secret
when user logs into existing gitlab account.