fix: Allow non-integer number format for score reporting (#168)
Current version of a checker returns score as a string that represents float number, which is casted directly to int. This cause a value error.

The proposed fis allow to push integers or floats in the score string, with the following behavior:

If integer is passed, it is taken as final score for the task
If a float less than zero is passed, set the score to 0
If a float larger than 2.0 is passed, report and error
If a float between 0.0 and 2.0 is passed, use it a fractional score, i.e. multiply it by the maximum score for the task.