feat: ask to re-type password (#156)
Currently, Manytask registration form asks to type password only once. Usually, when user is asked to set the password, it is typed twice. This is done to prevent setting password with a typo in it. Users are so used to having two forms when setting the password, they get confused whe asked to do it once, thinking that they need to type password that they set previously. This commit (1) adds another textfield to re-type password and (2) checks if two typed passwords are the same.