feat: create a database and make it possible to use it instead of gsheet (#194)
Although it is very convenient to use google spreadsheets as a database, there are certain limitations that come with it. First ad formost, the google sheet does not allow to make several courses on a single web-application instance. Secondly, this solution relies on google api and service accounts, which may become unavailable. This change is the first step to make it possible to use database. It makes all the data interfaces abstract, which allows to switch between two backends. Also this change contains the design of the database, which can be in future used to store the data on many courses. This change, however, keeps the one-web-application - one-course paradigm to keep backward compatibility and allow comparison testing.

### Commits

* feat: Create an abstract class for rating table

To make a seamless transition from gdoc api to database, we can use
an abstraction that represents the score table. In current version
of the code the RatingTable class can be used as such an abstraction.

* refactor: separate and make abstract gdoc api (#190)

* refactor: separate and make abstract GoogleDocApi

Remove fetch_rating_table method, instead of it transfer RatingTable methods to GoogleDocApi and transfer old GoogleDocApi methods to GoogleDocInit.

Create abstract classes: ViewerApi and StorageApi. Inherited GoogleDocApi from ViewerApi and StorageApi.

Change Course signature, instead of GoogleDocApi Course using ViewerApi and StorageApi.

* refactor: merge GoogleDocInit into GoogleDocApi

fix: remove unrelated changes and add dots to imports, because otherwise pytest doesn't work

* feat: add db models for DatabaseApi (#193)

feat: add db models for manytask web over  db

* fix: python imports sorting

---------

Co-authored-by: gagarinkomar <gagarinkomar@gmail.com>