Show the list of students names only when requested
Fetching the names one by one takes time, which delays the table
rendering. This adds and option to the request that enforces
showing names, but does not show them by default. When needed one
can add "show_full_names" oprion to the request which will result
in the table with names. Later can than be exported.