refactor: Move ownership of gitlab repo urls to Course class (#354)
* Remove GoogleSheet support

Google sheets do not allow for having many courses per Manytask
instance. This romves the support for the Google Sheets, making
the database sole storage choise.

* Move ownership of gitlab repo urls to Course class

In order to create forks for students, check if they have access and
simply provide urls to the tasks, we need to know where the course group,
public repo and students group are located and what is the name of the
default branch. This information was placed into the GitLab API class.
This was reasonable while one instance served one course: only one course
was handled by this API. To be able to reuse the API for many courses,
this information has to be taken out from the GitLab API class.

This change moves the ownership of the following data from the GitLabAPI
class into the Course class:
- gitlab_course_group - base group that encapsulates the students group and public repo
- gitlab_course_public_repo - public repo for the course
- gitlab_course_students_group - students group
- gitlab_default_branch - default branch

* Fix some tests

* Fix GitLab tests

* Fix tests

* Fix more function calls

* Fix formatting

* Add more typing annotationsin tests, define string in tests as a var

* Fix errors