Show percent of the accumulated score in the table
In most courses, the final grade is computed based on the
percent of the accumulated score. Thus it is convinient to
have this information in the table, rather than computing
it from the raw data. This chande computes the percent based
on the number of points in the currently enabled tasks.
Bonus tasks are not counted towards denominator. If the
total sum is zero, the percent is also shown as zero.

Closes issue #318