feat: remove link to the course in DataBaseApi
Removed storing course_name from DataBaseApi. Instead, passing course_name as a methods parameter. This removes the link to a single course and makes it possible to use multiple courses through one DataBaseApi.
Also removed course fields from DataBaseApi like registratin_secret etc...
Added method for course creation. It initializes the course and after that you can upload a config with deadlines to this course. We are currently initializing the course at startup, but in the future it will be easy to transfer the creation to the form.
Edcited tests.