Create RmsUser object and use it
Currently, the student object represent both RMS and logged user.
This can not be the case when the authorization and RMS logic wil
be separated. This change creates new object to store RMS user and
uses it thoughout the code. With that, the separation between
authorization and repository management becomes more visible.

Issue #395