refactor: Keep create user functionality in RMS (#417)
In some cases, it is valuable to disable creating users from RMS
directly (e.g. some additional verification is required for the
users). Hence, keeping an option to do it from Manytask  may become
useful. To simplify and not to polute the namespace, the User class
was removed in favour of string arguments. This class was created
specifically for this method and was used only here.

Issue #395