feat: Do not create user when store_score(..) and get/sync_stored_user(..) are called (#440)
Currently, in the guts of store_score(..), get_stored_user(..) and sync_stored_user(..) functions, the user is searched for and, if not found, is created. This should not be the case, because

- store_score(..) is user in API to set score(s) for task(s) and one should not try to set scores for non-existing user.
- get_stored_user(..) should not change anything - just return the user info
- sync_stored_user(..) should not create user in the system, only user on course

Also added tests that check that these functions throw an exception if called on non-existing user.

The function names are also confusing, which is in part why these functions are misused, see #449