feat: Add check_is_course_admin(..) method to StorageApi and use it (#453)
Currently, in the guts of store_score(..) function the user is searched for and, if not found, is created. This should not be the case, because one should not send scores for non-existing user via API. This also will cause an exception because there are no first and last name provided in the API call.