Create user when signing up, not when creating the first project
User on GitLab and in the database were created in two separate
places. Because of that we were parsing gitlab username into first
and last names. This moves the creation of the user in the database,
so it is created when the signup form is filled.

This is a follow-up for 75e4914ac35b6e427cb53e02a0ff52dd89e9f64b