feat:  refresh oauth token and return bool instead of throwing when checking authenticated RMS user (#497)
Refresh oauth token so that user will not have to re-login every 30 minutes

Checking if user logged into RMS should not throw and exception:
this is a routine check and nothing is wrong when user is not
authenticated. Hence this function should return boolean instead
of throwing an exception.

Changed the name of the function accordingly.

---------

Co-authored-by: Gorobets Ivan <96993672+KIoppert@users.noreply.github.com>