Set RMS ID for initial instance admin when app is initialized
When initial instance admin is created, their RMS ID is set to -1. This
causes an error when this user is identified later. Moreover, there is no
means to update the RMS ID from admin panel to fix this error. This fixes the
issue by getting RMS ID when instance is initialized.