feat: extra signup for new gitlab users (#547)
A separate page has been added for registering users who are accessing the system for the first time via GitLab.

Storing the username (probably better to use user_id in the future) in the session has been implemented.

Now, after logging in, the system checks whether the username is in the session. If it is not there, it checks whether the user is in the database. If the user is not found, they are redirected to a page to fill in additional fields.

The logic for redirecting to the page from which the user was sent to the login/signup page has been temporarily removed, as it significantly complicates things.