feat: Allow to use username in task url template (#615)
Currently, task template allows only group and task name macro. Hence the link to the task can not be to the students repo, for which one needs the student username. This adds macro for the student's username, which allows to link the student repo and make finding the task much more convinient.

---------

Co-authored-by: Artem Zhmurov <zhmurov@gmail.com>