docs: add structure to the TOC tree (#673)
This drafts the following sections into the docs:

Docs for students, including brief introduction to Git and general instruction on how to use Manytask (with links to onboarding course)
Docs for teachers on how to create course, what should be in the private repo and how to set up deployment
Docs for admins so that they can deploy and maintain their own instance of Manytask
Docs for developers on how to start contributing to the code, run the dev instance and test the code locally.
Docs from checker repo should also be updated and transferred to the new location.
---------

Co-authored-by: Aleksei Seliverstov <142227259+prawwtocol@users.noreply.github.com>