docs: Remove old development strategy, while keeping course as code description (#677)
The development strategy was composed for the previous iteration, and
most of the tasks listed there are now completed. Hence there is no
need in keepin this plan. The course as code section is the main idea
for the development, and it should be kept in the teacher's section of
the docs.