feat: Create in-memory RMS and use it in tests (#535)
This creates an implementation of the RMS API class, that mimics the behavior of the RMS. It stores users and projects in memory, which makes it useful in tests (to mock the RMS API) and while developing (if one does not have access to RMS). By default, GitLab will be used as an RMS, but one can change this by setting RMS envvar to "mock" in the .env file.