Allow reducting students score with api (#734)
Currently, if the reported score is lower than the score in the system,
max value is taken. This adds 'allow_reduction' parameter to the API,
which allows for reducing the score for the student. The default value
is False, so that the bahavior does not change for the current code.